チャット画面オファータブ追加
