+++ components/chat/chat.style.ts
@@ -2,11 +2,8 @@ import { makeStyles } from "@material-ui/core";
 
 export const useStyles = makeStyles((theme) => ({
 
-    header: {
-        zIndex: 600,
-        position: "fixed",
-        width: theme.spacing(200),
-        backgroundColor: "var(--color-white)",
+    offerTab: {
+        backgroundColor: "green",
     },
 
     haderContent: {
@@ -15,6 +12,7 @@ export const useStyles = makeStyles((theme) => ({
 
     headerLinks: {
         float: "right",
+        // textAlign:"right",
         backgroundColor: "rgba(253, 255, 130, 0.755)",
     },
  }));
