+++ components/chat/components/chat-tab/chat-tab.module.scss
@@ -4,7 +4,6 @@
     width: 500px;
 }
 
-
 .kaiwaImgLeft {
     width: 14%;
     display: inline-block;
