+++ components/header/header.module.scss
@@ -3,7 +3,7 @@
     position: fixed;
     width: 100%;
     background-color: white;
-    box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
+    // box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
     box-sizing: border-box;
 }
 
@@ -18,7 +18,6 @@
 .searchFieldInput {
     margin: 8px;
     width: 30%;
-    // width: 500px;
 }
 
 .headerLinks {
